Use addEventListener. Behavior changes.
When a route is not found during AJAX navigation, fall back to server
rendering. Do this because the server may wish to deliver a static resource, or
a page NOT in the Pages routing table, etc. Pages was getting in the way by
rendering an assumed-to-exist 404 route on the client-side (the possible
nonexistence of a 404 route causes other issues anyway).